==> onsides_intl/onsides_jp/notebooks/3-0.term_extraction_ade.ipynb <==
import numpy as np
import pandas as pd
import requests
from tqdm import tqdm
from glob import glob
import ast, json
from bs4 import BeautifulSoup
import warnings
warnings.filterwarnings('ignore')
#specify home folder (all sub-folders, scripts, data in this folder)
data_folder = ''
api = ''
date = ''

def format_code(code):
    #this function takes in the code, and prepend zeros to the left of the string to make it all 8 characters.
    formatted_code = str(code).zfill(8)
    return formatted_code

##read in table for all rx drugs
kegg_df = pd.DataFrame()
for f in glob(data_folder+'data/kegg_rx_drug_data*'):
  kegg_df = pd.concat([kegg_df, pd.read_csv(f)], axis = 0)
  print(f, kegg_df.shape[0])
kegg_df = kegg_df.drop_duplicates()
kegg_df['japic_code'] = kegg_df['kegg_product_id'].apply(format_code)
print(kegg_df.shape[0])
kegg_df.head(1)

ades_df = pd.read_csv(data_folder+'data/rx_drug_ade_raw_all_v{}.csv'.format(date))
ades_df['japic_code'] = ades_df['japic_code'].apply(format_code)
ades_df = ades_df.drop_duplicates()
print(ades_df.shape, len(ades_df.japic_code.unique().tolist()))
ades_df.head(1)

---

#extract only the text
text_strings = []
for i, row in tqdm(ades_df.iterrows()):
  string = BeautifulSoup(row['ade'], 'html.parser')
  if len(string.find_all('table')) > 0:
    string.table.decompose()
    string = string.text.strip()
  else:
    string = None
  text_strings.append(string)
ades_df['ade_text'] = text_strings
ades_df.head(1)

#unique strings
ades_strings = ades_df[ades_df.ade_text.notna()][['ade_text']].drop_duplicates()
ades_strings.to_csv(data_folder+'data/ade/rx_drug_ade_text_v{}.csv'.format(date), index = False)

---

#extract only the table
tables = pd.DataFrame()
table_strings = []
table_strings_failed = []
for i, row in tqdm(ades_df.iterrows()):
  string = BeautifulSoup(row['ade'], 'html.parser')
  if len(string.find_all('table')) > 0:
    try:
      table = string.find_all('table')[0]
      ades = pd.read_html(str(table))[0]
      ades.columns = ades.iloc[0]
      ades = ades.rename(columns={np.nan: 'category'}).drop(ades.index[0]).reset_index()
      ades['japic_code'] = row['japic_code']
      tables = pd.concat([tables, ades], axis = 0)
    except:
      table_strings_failed.append(row['japic_code'], string.find_all('table')[0])
  else:
    table = None

tables.head(1)
