==> onsides_intl/onsides_jp/notebooks/5.format_tables.ipynb <==
import numpy as np 
import pandas as pd
import requests
from tqdm import tqdm
import ast, json
from bs4 import BeautifulSoup
import warnings
warnings.filterwarnings('ignore')
data_folder = ''
external_data_folder = ''
api = ''

- ingredient_rxcuis, string list of comma separated RxNORM identifiers for the ingredients
- ingredient_names, string list of comma separated ingredients in the drug product (most are single but many combination medications as well)
- num_ingredients, number of ingredients in this drug product
- pt_meddra_id, the MedDRA preferred term code for the adverse reaction term identified
- pt_meddra_term, the MedDRA preferred term for the adverse reaction term identified
- pt_meddra_jp_term, the MedDRA/J preferred term for the adverse reaction term identified 
- percent_labels, the proportion of labels for which this adverse reaction was extracted (0-1)
- num_labels, the number of active labels for this ingredient or ingredient combination

kegg_df = pd.read_csv(data_folder+'kegg_rx_drug_data.csv')
kegg_df.columns = ['product', 'ingredient', 'indication', 'kegg_id', 'japic_code']
condition_df = pd.read_csv(data_folder+'ade/rx_ade_table.csv')
conditions = condition_df[['japic_code', 'ade', 'category', 'percentage']]
conditions = conditions.merge(kegg_df[['japic_code', 'kegg_id']], on = 'japic_code', how = 'left')
unique_ades = pd.read_csv(data_folder+'ade/rx_raw_ade_list.csv')
unique_ades['SDUI'] = unique_ades.SDUI.apply(lambda x: ast.literal_eval(x) if str(x)!='nan' else None)
unique_ades = unique_ades[['ade', 'SDUI']].explode('SDUI')
conditions = conditions.merge(unique_ades[['ade', 'SDUI']], on = 'ade', how = 'left')
conditions = conditions[['japic_code', 'percentage', 'SDUI']]
conditions = conditions[conditions.SDUI.notna()]
conditions.japic_code = conditions.japic_code.apply(lambda x: str(int(x)))
print(conditions.shape)
conditions.head()

## Map Drgs

drug_df = pd.read_csv(data_folder+'rx_drug_overview_parsed.csv')
drug_df['KEGG DRUG'] = drug_df['KEGG DRUG'].apply(lambda x: BeautifulSoup(x) if str(x) != 'nan' else None)
drug_df['KEGG DRUG'] = drug_df['KEGG DRUG'].apply(lambda x: [i.text for i in x.find_all('a') if i.text[0] == 'D'][0] if x != None else None)
drug_df['en_name'] = drug_df['欧文一般名'].apply(lambda x: x.lower() if str(x) != 'nan' else None)
drug_df.head()

drug_filtered = drug_df[['japic_code', 'version', '総称名', '薬効分類名', 'ATCコード', 'en_name']]
drug_filtered.to_csv(data_folder+'final/drug_table.csv', index=False)
drug_filtered.head(1)

rxnorm = pd.read_csv(external_data_folder+'umls_rxnorm.csv')
rxnorm = rxnorm[['CODE', 'STR']]
rxnorm.STR = rxnorm.STR.apply(lambda x: x.lower())
rxnorm.head(1)

drug_df = drug_df.merge(rxnorm, left_on = 'en_name', right_on = 'STR', how = 'left')
drug_df = drug_df[['japic_code', '一般名', 'en_name', 'KEGG DRUG', 'CODE']]
drug_df['KEGG DRUG'] = drug_df['KEGG DRUG'].apply(lambda x: [i.text for i in x.find_all('a') if i.text[0] == 'D'][0] if x != None else None)
drug_df.head(1)

kegg_map = pd.read_csv(data_folder+'drug_atc_kegg.csv')
kegg_dict = dict(zip(kegg_map.kegg, kegg_map.rxnorm))
drug_df['CODE'] = drug_df.apply(lambda x: kegg_dict(x['KEGG DRUG']) if x.CODE == None and x['KEGG DRUG'] in kegg_dict.keys() else x.CODE, axis = 1)
kegg_map.head(1)

drug_df.notna().sum()

df = pd.read_csv(data_folder+'kegg_drug_info_mapped.csv')
df.head(1)

drug_df = pd.read_csv(data_folder+'drug_atc_kegg.csv')
drug_df = drug_df.merge(kegg_df, left_on = 'kegg', right_on = 'kegg_id', how = 'left')
drug_df = drug_df[drug_df.japic_code.notna()]
drug_df['japic_code'] = drug_df.japic_code.apply(lambda x: str(int(x)))
drug_df.head(1)

## Map to Raw

umls_rxnorm = pd.read_csv(external_data_folder+'umls_rxnorm.csv')[['CODE', 'STR']].drop_duplicates(subset='CODE')
umls_rxnorm.columns = ['ingredients_rxcuis', 'ingredient_names']
umls_meddra = pd.read_csv(external_data_folder+'umls_meddra_en.csv')[['CODE', 'STR']].drop_duplicates(subset='CODE')
umls_meddra.columns = ['pt_meddra_id', 'pt_meddra_term']

raw_tally = conditions.merge(drug_df[['japic_code','rxnorm']], on = 'japic_code', how = 'left')\
.merge(umls_rxnorm, left_on = 'rxnorm', right_on = 'ingredients_rxcuis', how = 'left')\
.merge(umls_meddra, left_on = 'SDUI', right_on = 'pt_meddra_id', how = 'left')
raw_tally = raw_tally.drop(['SDUI', 'rxnorm', 'percentage'], axis = 1)
raw_tally.to_csv(data_folder+'final/onsides_raw_table_wo_freq.csv', index=False)
raw_tally.head(1)

onsides_df = pd.DataFrame()
for ingredient in tqdm(drug_df.rxnorm.unique().tolist()):
  product_list = drug_df[drug_df.rxnorm == ingredient].japic_code.tolist() 
  num_labels = len(product_list)
  filtered_conditions = conditions[conditions.japic_code.isin([i for i in product_list])].drop_duplicates()
  individual_df = filtered_conditions.SDUI.value_counts().reset_index()
  individual_df.columns = ['pt_meddra_id', 'percent_labels']
  individual_df['ingredients_rxcuis'] = ingredient
  individual_df['num_labels'] = num_labels
  individual_df['percent_labels'] = individual_df.percent_labels.apply(lambda x: x / num_labels)
  onsides_df = onsides_df.append(individual_df)
onsides_df.head()

onsides_df = onsides_df.merge(umls_rxnorm, on = 'ingredients_rxcuis', how = 'left')\
.merge(umls_meddra, on = 'pt_meddra_id', how = 'left')
onsides_df = onsides_df[['ingredients_rxcuis', 'ingredient_names', 'pt_meddra_id', 'pt_meddra_term', 'percent_labels', 'num_labels']]
print(onsides_df.shape)
onsides_df.head(1)

onsides_df.to_csv(data_folder+'final/onsides_compiled_table_wo_freq.csv', index=False)
